+++ pkg/super_node/btc/payload_fetcher.go
@@ -55,7 +55,7 @@ func (fetcher *PayloadFetcher) FetchAt(blockHeights []uint64) ([]shared.RawChain
 			return nil, err
 		}
 		blockPayloads[i] = BlockPayload{
-			Height: int32(height),
+			Height: int64(height),
 			Header: &block.Header,
 			Txs:    msgTxsToUtilTxs(block.Transactions),
 		}
