+++ pkg/super_node/btc/streamer.go
@@ -49,7 +49,7 @@ func (ps *PayloadStreamer) Stream(payloadChan chan shared.RawChainData) (shared.
 		// Notification handler for block connections, forwards new block data to the payloadChan
 		OnFilteredBlockConnected: func(height int32, header *wire.BlockHeader, txs []*btcutil.Tx) {
 			payloadChan <- BlockPayload{
-				Height: height,
+				Height: int64(height),
 				Header: header,
 				Txs:    txs,
 			}
